requirements.txt
numpy
matplotlib
traci
torch
kafka-python
